--- src/helpers.ts
@@ -1,14 +1,0 @@
-export const mapToObject = (map: Map<string, any>): { [k: string]: any } =>
-    Object.fromEntries(map.entries());
-
-    export const makeid = (length: number = 10): string => {
-        let result = '';
-        const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-        const charactersLength = characters.length;
-        let counter = 0;
-        while (counter < length) {
-            result += characters.charAt(Math.floor(Math.random() * charactersLength));
-            counter += 1;
-        }
-        return result;
-    }
