+++ src/node/RootNode.ts
@@ -1,4 +1,4 @@
-import { mapToObject } from "../helpers";
+import { mapToObject } from "../utils/helpers";
 import { XsNode } from "./XsNode";
 
 export class RootNode extends XsNode {
